Add files via upload
Overview of the statistical analysis:
New York Citibike is a city public program. The program is a public transportation option for New Yorkers with bike storage locations around the city, easy registration, and easy rental and return. New York Citibike is also a New York City experience program for visitors from around the world, providing visitors with a new experience and an alternative to the city scenery. The operation mode of this project in New York has been very perfect and is in the stage of continuous profit. We will explore whether this mode of operation is replicable and sustainable.

Checkout Time vs Users:
The majority of Newyork Citibike users take short trips, between 1 and 3 minutes. There are transit stations between buses and subways for people going to work, or distances for students to travel from dormitories to school.

Checkout time by Gender:
Males make up a very large proportion of rental excursions.

Trips by weekday for each hour:
Weekdays are 6-9 a.m. and 5-7 p.m. are peak times for the citibike rentals in New York City. This project has effectively eased the traffic in New York City to some extent, and some users who need to travel short and medium distances no longer need to take a car.

Trips by gender:
Males make up the majority of users on both weekdays and weekends. So Citibike storage points can be set up in areas with a large proportion of male employees.

Trips by gender by weekdays:
Male users make up more subscribers and have a stronger preference for public transport.

Summary:
New York Citibike is a very successful public transportation program, but it doesn't work in all cities. The business model is replicable, but profitability is not guaranteed because user ratios are wildly skewed, with men having a stronger preference for subscriptions. The main use area of New York Citibike should be the area where office buildings and school dormitories are located, providing users with short and medium trips. If the range is too large and too scattered, the opportunity cost of storing the bikes will be very high, and the operating cost will increase accordingly.